pyproject.toml
[tool.poetry]
name = "download-bgt"
version = "0.1.0"
description = ""
authors = ["Davide Motta <davide.motta@luxottica.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^2.2.3"
PyYAML = "^6.0.2"
customtkinter = "^5.2.2"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
